Bias_Detection_Agent/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.8.0
pyyaml>=6.0
streamlit>=1.33.0
streamlit-option-menu>=0.3.0
seaborn>=0.12.0
plotly>=5.0.0
groq>=0.4.0
